Make AbstractPerformanceTest support dynamic rate control.
Provide methods in AbstractPerformanceTest that can be overridden to control the
rate limiting. This is useful if the target rate the performance test wants to
hit cannot be expressed as a constant rate of iterations per second, e.g. if it
actually wants messages/s or bytes/s and each iteration has a variable number of
messages or bytes.