Add a structured logger and log4j1/2 extensions for logging to kafka (#173)
* Add a structured logger and log4j2 extensions for logging to kafka

This patch adds interfaces for structured logging. Structured log messages
are represented as instances of connect SchemaAndValue, and would typically
contain objects of type Struct. The main interfaces to the structured logger
are a factory and the logger itself. The factory takes a prefix to prepend
to the name of each logger it creates.

The logger has functions for logging structured messages at various log levels.
The implementation just logs to slf4j using the parameters
<level>.("{}", structuredMsg). The underlying logging implementation can be
extended to handle structured msgs in more interesting ways (e.g. serializing
them as avro/json and writing to kafka).

The object passed to slf4j is of type SerializableSchemaAndValue, which
adds a toString message that returns the value as json. This way, the
messages are printed to the log (e.g. a file or console) in json by default.

The logger also has interfaces that accept Supplier<SchemaAndValue>, so users
can avoid constructing messages if the logger is disabled.

This patch also includes a module with some log4j2 extensions for formatting
structured messages as json or avro. This can be used in conjunction with the
KafkaAppender (which ships with log4j2) to log structured messages to kafka.

* some clean-up plus log4j1 impl

* Forgot to add the struct builder test

* Address review feedback

* Add missing pom

* clean up guava from root pom

* Fix review feedback

* typo

* typo

* Clean up references to topic

* review feedback