Upgrade to ZooKeeper 3.5.5 and remove zkclient dependency (#211)
Apache Kafka has upgraded to ZooKeeper 3.5.5 recently:

https://github.com/apache/kafka/pull/6802

Kafka has also removed `ZkUtils` and the `zkclient` dependency. We do the
same here (remove the `ZkUtils` clone and the `zkclient` dependency). We
keep the `dependencyManagement` entries for `zookeeper` and `zkclient` as
other Confluent projects have dependencies on those libraries.

ZooKeeper 3.5.5 is the first stable release in the 3.5.x series. The key new feature
is TLS support, but there are a few more noteworthy features:

* Dynamic reconfiguration
* Local sessions
* New node types: Container, TTL
* Ability to remove watchers
* Multi-threaded commit processor
* Upgraded to Netty 4.1

See the release notes for more detail:
https://zookeeper.apache.org/doc/r3.5.5/releasenotes.html