WIP - *DO NOT MERGE* - Example gh-actions cross-platform build
A 7.0.0 cross-platform OCI-compatible cp-base-new docker-image for
the `amd64/x86_64` and `arm64` architectures. This is
a port of the POC repo I have at
[confluent-cp-images](https://github.com/jackcviers/confluent-cp-images)
as a PR for this repository.

The intent of this pr is demonstration of concept. If Confluent
chooses to adopt this for their public images, we can change the
copyright/licenses etc.

The GH Actions will require the definition of the following repository secrets:

DOCKER_HUB_USERNAME - the username used for publishing to the docker
registry
DOCKER_HUB_ACCESS_TOKEN - the access token used for publishing to the
docker registry

The DOCKER_HUB_USERNAME secret is used to define ${REGISTRY_USER} in
the makefile, which is set to `unknown` by default, which should
prevent us from clobbering anything else named cp-base-new locally.

The build is confirmed to work on macOS Monterrey on Apple M1 silicon,
macOS Big Sur on Intel, Ubuntu 20.04 AMD64, and on the Azure GH
actions ubuntu-20.04 instances.

Please see the README in this pr for local build pre-requisites, under
`Build.Prerequisites.<PLATFORM>`.